Refactor dashboard components
